codebook inference no long blocks 256+ categories

+++ lace/lace_codebook/src/data.rs
@@ -335,7 +335,7 @@ fn string_categorical_coltype(
     use std::collections::BTreeSet;
 
     let n_unique = srs.n_unique()?;
-    if n_unique > std::u8::MAX as usize {
+    if n_unique > std::u32::MAX as usize {
         Err(CodebookError::CategoricalOverflow {
             col_name: srs.name().to_owned(),
         })
